requirements.txt
altgraph==0.14
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
coverage==4.4.1
Django==1.11.3
django-constance==2.0.0
django-debug-toolbar==1.8
django-filter==1.0.4
django-oauth-toolkit==1.0.0
django-picklefield==1.0.0
django-rest-swagger==2.1.2
djangorestframework==3.6.3
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
macholib==1.8
MarkupSafe==1.0
modulegraph==0.15
oauthlib==2.0.2
openapi-codec==1.3.2
pytz==2017.2
requests==2.18.1
simplejson==3.11.1
sqlparse==0.2.3
uritemplate==3.0.0
urllib3==1.21.1
